Updated clubsandwich-0.1.3-py3-none-any.whl to have the scrollingtextview and moved forerunner out of game scene.
